+++ src/index.browser.ts
@@ -1,1 +1,9 @@
-export {}
+export * from './index.browser.js'
+export * from './Placement.js'
+export * from './set_svg_matrix2d_b.js'
+export * from './svg_content_height$_b.js'
+export * from './svg_content_width$_b.js'
+export * from './svg_height$_b.js'
+export * from './svg_margin$_b.js'
+export * from './svg_matrix2d_ctx$_b.js'
+export * from './svg_width$_b.js'