+++ src/markup.ts
@@ -1,3 +1,0 @@
-import { markup_, markup_T } from './markup_.js'
-export const markup:markup_T = markup_()
-export { markup as markup__markdown}