+++ src/svg_content_width$_b.ts
@@ -1,10 +1,10 @@
+import { WritableAtom$, atom$ } from '@ctx-core/nanostores'
 import { B, be_ } from '@ctx-core/object'
-import { Writable$, writable$ } from '@ctx-core/store'
 const key = 'svg_content_width$'
 export const svg_content_width$_b:B<svg_content_width$_T> = be_(key, ()=>
-	writable$(undefined) as svg_content_width$_T
+	atom$(undefined) as svg_content_width$_T
 )
-export type svg_content_width$_T = Writable$<number|undefined>
+export type svg_content_width$_T = WritableAtom$<number|undefined>
 export {
 	svg_content_width$_b as b__width__content__svg,
 }