+++ src/svg_matrix2d_ctx$_b.ts
@@ -1,27 +1,27 @@
 import { B, be_ } from '@ctx-core/object'
-import { derived, Readable } from '@ctx-core/store'
+import { computed$, ReadableAtom$ } from '@ctx-core/nanostores'
 import { svg_margin$_b } from './svg_margin$_b.js'
 import { svg_width$_b } from './svg_width$_b.js'
 import { svg_height$_b } from './svg_height$_b.js'
 import { svg_content_width$_b } from './svg_content_width$_b.js'
 import { svg_content_height$_b } from './svg_content_height$_b.js'
-import type { Placement } from './Placement'
+import type { Placement } from './Placement.js'
 const key = 'svg_matrix2d_ctx$'
 export const svg_matrix2d_ctx$_b:B<svg_matrix2d_ctx$_T> = be_(key, ctx=>
-	derived([
+	computed$([
 			svg_margin$_b(ctx),
 			svg_width$_b(ctx),
 			svg_height$_b(ctx),
 			svg_content_width$_b(ctx),
 			svg_content_height$_b(ctx),
 		],
-		([
-			 svg_margin,
-			 svg_width,
-			 svg_height,
-			 svg_content_width,
-			 svg_content_height
-		 ])=>({
+		(
+			svg_margin,
+			svg_width,
+			svg_height,
+			svg_content_width,
+			svg_content_height
+		)=>({
 			svg_margin,
 			svg_width,
 			svg_height,
@@ -37,7 +37,7 @@ export interface svg_matrix2d_ctx_T {
 	svg_content_width:number|undefined
 	svg_content_height:number|undefined
 }
-export type svg_matrix2d_ctx$_T = Readable<svg_matrix2d_ctx_T>
+export type svg_matrix2d_ctx$_T = ReadableAtom$<svg_matrix2d_ctx_T>
 export {
 	svg_matrix2d_ctx$_b as b__ctx__matrix2d__svg,
 }