+++ src/svg_width$_b.ts
@@ -1,10 +1,10 @@
 import { B, be_ } from '@ctx-core/object'
-import { writable$, Writable$ } from '@ctx-core/store'
+import { atom$, WritableAtom$ } from '@ctx-core/nanostores'
 const key = 'svg_width$'
 export const svg_width$_b:B<svg_width$_T> = be_(key, ()=>
-	writable$(undefined) as svg_width$_T
+	atom$(undefined) as svg_width$_T
 )
-export type svg_width$_T = Writable$<number|undefined>
+export type svg_width$_T = WritableAtom$<number|undefined>
 export {
 	svg_width$_b as b__width__svg,
 }